% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{bm_25}
\alias{bm_25}
\alias{bm_25.dtm}
\alias{bm_25.corpus}
\title{Okapi BM-25}
\usage{
bm_25(text)

\method{bm_25}{dtm}(text)

\method{bm_25}{corpus}(text)
}
\arguments{
\item{text}{A document-term martrix or a \code{corpus}.}
}
\description{
Okapi BM25 (BM stands for Best Matching) is a ranking 
function used by search engines to estimate the relevance 
of documents to a given search query.
}
\examples{
\dontrun{
init_textanalysis()

# create corpus
doc <- string_document("A simple document.")
doc2 <- string_document("Another simple document.")
crps <- corpus(doc, doc2)

# matrix & plot
bm_25(crps)
}

}
