% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartBigrams.R
\name{chartBigrams}
\alias{chartBigrams}
\title{Chart bigrams}
\usage{
chartBigrams(bigrams)
}
\arguments{
\item{bigrams}{df containing just word1 and word2}
}
\description{
Chart bigrams
}
\examples{
df<-data.frame(word1=LETTERS[1:5], word2=LETTERS[2:6])
chartBigrams(df)

}
