% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{corpus}
\alias{corpus}
\title{Corpus}
\usage{
corpus(...)
}
\arguments{
\item{...}{Documents.}
}
\description{
Build a corpus of documents.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc1 <- string_document("First document.")
doc2 <- string_document("Second document.")

corpus <- corpus(doc1, doc2)
}

}
