% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{corpus_to_tibble}
\alias{corpus_to_tibble}
\alias{corpus_to_tibble.corpus}
\title{Convert Corpus}
\usage{
corpus_to_tibble(corpus)

\method{corpus_to_tibble}{corpus}(corpus)
}
\description{
Convert a corpus to a \link[tibble]{tibble}.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc1 <- string_document("First document.")
doc2 <- string_document("Second document.")

corpus <- corpus(doc1, doc2)

corpus_to_tibble(corpus)
}

}
