% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_DTM.R
\name{create_DTM}
\alias{create_DTM}
\title{Create a tidy document-term data_frame}
\usage{
create_DTM(df, ID, text, n_gram = 1, stop_rm = TRUE, stemming = FALSE,
  q = NULL)
}
\arguments{
\item{df}{The original data frame (LexisNexis, Twitter, etc)}

\item{ID}{The name of the variable identifying each document in the original data frame}

\item{text}{The name of the variable containing the texts we want to analyze in the original data frame}

\item{n_gram}{A numeric specifying the the number of grams. If default to NULL, then use 1-gram}

\item{stemming}{A logical specifying whether to implement stemming. Default to FALSE}

\item{q}{A numeric specifying the quantile of tf-idf to remove words. If default to NULL, then don't remove}
}
\value{
A tidy document-term data_frame containing ID, word, word counts, and tf-idf
}
\description{
Throw in the original data frame containing text, extract the text,
then create a document-term data_frame.
}
\examples{
create_DTM(document, ID, FULL_TEXT)
create_DTM(document, ID, FULL_TEXT, 1, TRUE, 0.2)
}
\author{
Jiacheng He
}
