% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document_metadata}
\alias{document_metadata}
\alias{title_}
\alias{title_.document}
\alias{language_}
\alias{language_.document}
\alias{author_}
\alias{author_.document}
\alias{timestamp_}
\alias{timestamp_.document}
\title{Metadata}
\usage{
title_(document, ...)

\method{title_}{document}(document, ...)

language_(document, ...)

\method{language_}{document}(document, ...)

author_(document, ...)

\method{author_}{document}(document, ...)

timestamp_(document, ...)

\method{timestamp_}{document}(document, ...)
}
\arguments{
\item{document}{A document as returned by the \code{*_document} 
family of functions, i.e.: \code{\link{string_document}}.}

\item{...}{A character string to set.}
}
\description{
Get and set document metadata.
}
\details{
Note that the value (three dots) for the \code{language}
function is taken from the Julia \code{Language} package.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc <- string_document("This is a document.")

# get and et metadata
title_(doc) # get
title_(doc, "Hello World") # set
language_(doc, "Spanish") # from Language pack
}

}
