% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{document_term_matrix}
\alias{document_term_matrix}
\alias{document_term_matrix.corpus}
\title{DocumentTermMatrix}
\usage{
document_term_matrix(corpus)

\method{document_term_matrix}{corpus}(corpus)
}
\arguments{
\item{corpus}{A corpus, as returned vy \code{\link{corpus}}.}
}
\description{
Represent documents as a matrix of word counts.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc1 <- string_document("First document.")
doc2 <- string_document("Second document.")

corpus <- corpus(doc1, doc2)

update_lexicon(corpus)
m <- document_term_matrix(corpus)
}

}
