% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{document_term_vector}
\alias{document_term_vector}
\alias{document_term_vector.corpus}
\title{Document Term Vector}
\usage{
document_term_vector(corpus, document)

\method{document_term_vector}{corpus}(corpus, document)
}
\arguments{
\item{corpus}{A corpus, as returned vy \code{\link{corpus}}.}

\item{document}{Index of document within the corpus.}
}
\description{
Produce a single row of a DocumentTermMatrix.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc1 <- string_document("First document.")
doc2 <- string_document("Second document.")

corpus <- corpus(doc1, doc2)

update_lexicon(corpus)
document_term_vector(corpus)
}

}
