% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{documents}
\alias{documents}
\alias{file_document}
\alias{string_document}
\alias{token_document}
\alias{ngram_document}
\alias{as_documents}
\title{Document}
\usage{
file_document(path)

string_document(text)

token_document(text)

ngram_document(text, ...)

as_documents(text, type = c("string", "token", "ngram"))
}
\arguments{
\item{path}{The path to the file.}

\item{text}{The text, tokens, or ngrams.}

\item{...}{Other positonal arguments.}

\item{type}{Type of \code{*_document} function to use.}
}
\description{
The basic unit of text analysis is a document. 
The textanalysis package allows one to work with documents 
stored in a variety of formats.
}
\examples{
\dontrun{
init_textanalysis()

doc <- "This is a document."
string_document(doc)
ngram_document(doc, 2L)
}

}
\seealso{
\code{\link{directory_corpus}} to read a dicrectory of files as corpus.
}
