% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{dtm_matrix}
\alias{dtm_matrix}
\alias{dtm_matrix.dtm}
\title{Sparse Matrix}
\usage{
dtm_matrix(dtm, density)

\method{dtm_matrix}{dtm}(dtm, density = c("sparse", "dense"))
}
\arguments{
\item{dtm}{An object of class \code{document_term_matrix} 
as returned by \code{document_term_matrix}.}
}
\description{
Creates a simple sparse matrix of \code{document_term_matrix} object.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc1 <- string_document("First document.")
doc2 <- string_document("Second document.")

corpus <- corpus(doc1, doc2)

update_lexicon(corpus)
m <- document_term_matrix(corpus)
dtm_matrix(m) # sparse
dtm_matrix(m, "dense") # dense
}

}
