% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Sentiment.R
\name{extract_sentiment}
\alias{extract_sentiment}
\title{Extract sentiment vectors from a document-term data_frame}
\usage{
extract_sentiment(df, lexicon)
}
\arguments{
\item{df}{The document-term data_frame output from create_DTM}

\item{lexicon}{The sentiment database. lexicon = c("affin", "bing", "nrc", "loughran")}
}
\value{
A data_frame where each row represents a document and each column represents a sentiment.
If we use "affin" then it returns the sentiment score
}
\description{
Input the document-term data_frame, match each word in each document
to a sentiment database, and calculate the sentiment intensity of each document
}
\examples{
extract_sentiment(text, "nrc")
extract_sentiment(text, "affin")
}
\author{
Jiacheng He
}
