% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{get_ngrams}
\alias{get_ngrams}
\alias{get_ngrams.document}
\title{Extract NGrams}
\usage{
get_ngrams(document, ...)

\method{get_ngrams}{document}(document, ...)
}
\arguments{
\item{document}{A document as returned by the \code{*_document} 
family of functions, i.e.: \code{\link{string_document}}.}

\item{...}{Any other positional arguments.}
}
\value{
A \link[tibble]{tibble} of ngrams and their occurences.
}
\description{
Access n-grams tokens of documents as a vector.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc <- string_document("This is a document.")

# extract text
get_ngrams(doc)
get_ngrams(doc, 2L)
}

}
