% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{get_tokens}
\alias{get_tokens}
\alias{get_tokens.document}
\alias{get_tokens.JuliaObject}
\title{Extract Tokens}
\usage{
get_tokens(document)

\method{get_tokens}{document}(document)

\method{get_tokens}{JuliaObject}(document)
}
\arguments{
\item{document}{A document as returned by the \code{*_document} 
family of functions, i.e.: \code{\link{string_document}}.}
}
\description{
Access tokens of documents as a vector.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc <- string_document("This is a document.")

# extract tokens
get_tokens(doc)
}

}
