% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScript.R
\name{getScript}
\alias{getScript}
\title{Return a script for a given URL}
\usage{
getScript(df)
}
\arguments{
\item{df}{Data,frame}
}
\value{
a character vector
}
\description{
Return a script for a given URL
}
\examples{
df<-getScriptURLs()
getScript(df[1,])
}
