% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScriptAllBigrams.R
\name{getScriptAllBigrams}
\alias{getScriptAllBigrams}
\title{Get bi-grams used in Monty Python movie lines}
\usage{
getScriptAllBigrams(offline = FALSE)
}
\arguments{
\item{offline}{Use an offline copy instead of fetching data}
}
\value{
data.frame of bi-grams from Monty Python scripts
}
\description{
Get bi-grams used in Monty Python movie lines
}
\examples{
head(getScriptAllBigrams(offline=TRUE))
}
