% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScriptAllWords.R
\name{getScriptAllWords}
\alias{getScriptAllWords}
\title{Get words used in Monty Python movie lines, without stop words removed}
\usage{
getScriptAllWords(offline = FALSE)
}
\arguments{
\item{offline}{Use an offline copy instead of fetching data}
}
\value{
data.frame of words from Monty Python scripts
}
\description{
Get words used in Monty Python movie lines, without stop words removed
}
\examples{
head(getScriptAllWords(offline=TRUE))
}
