% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScriptBigrams.R
\name{getScriptBigrams}
\alias{getScriptBigrams}
\title{Get bi-grams used in Monty Python movie lines, without stop words}
\usage{
getScriptBigrams(offline = FALSE)
}
\arguments{
\item{offline}{Use an offline copy instead of fetching data}
}
\value{
data.frame of bi-grams from Monty Python scripts
}
\description{
Get bi-grams used in Monty Python movie lines, without stop words
}
\examples{
head(getScriptBigrams(offline=TRUE))
}
