% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScriptLines.R
\name{getScriptLines}
\alias{getScriptLines}
\title{Get Monty Python script lines}
\usage{
getScriptLines(offline = FALSE)
}
\arguments{
\item{offline}{Use an offline copy instead of fetching data}
}
\value{
data.frame of lines from Monty Python scripts
}
\description{
Get Monty Python script lines
}
\examples{
head(getScriptLines(offline=TRUE))
}
