% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScriptSpeech.R
\name{getScriptSpeech}
\alias{getScriptSpeech}
\title{Get Monty Python script lines that are speech}
\usage{
getScriptSpeech(offline = FALSE)
}
\arguments{
\item{offline}{Use an offline copy instead of fetching data}
}
\value{
data.frame of lines from Monty Python scripts
}
\description{
Get Monty Python script lines that are speech
}
\examples{
head(getScriptSpeech(offline=TRUE))
}
