% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScriptURLs.R
\name{getScriptURLs}
\alias{getScriptURLs}
\title{Get Monty Python script URLs}
\usage{
getScriptURLs()
}
\value{
tibble containing script name and URL
}
\description{
Based on \code{MONTYPYTHONSCRIPTSITE}, retrieve relevant
URLs of Monty Python scripts.
}
\examples{
df<-getScriptURLs()
head(df)
}
