% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScriptWords.R
\name{getScriptWords}
\alias{getScriptWords}
\title{Get words used in Monty Python movie lines, with stop words removed}
\usage{
getScriptWords(offline = FALSE)
}
\arguments{
\item{offline}{Use an offline copy instead of fetching data}
}
\value{
data.frame of words from Monty Python scripts
}
\description{
Get words used in Monty Python movie lines, with stop words removed
}
\examples{
head(getScriptWords(offline=TRUE))
}
