% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSentiment.R
\name{getSentiment}
\alias{getSentiment}
\title{Send text to Microsoft Cognitive Services' Sentiment API}
\usage{
getSentiment(textdf, apikey = NULL)
}
\arguments{
\item{textdf}{A data.frame consisting of two cols with colnames \code{c("id","text")}.
Optionally you can also provide a "language" column with ISO country codes,
otherwise it will default to "en".}

\item{apikey}{Your key for working with Microsoft Cognitive Services}
}
\value{
response A data.frame with id and a sentiment score
}
\description{
Send lines of text to an API to get the sentiment score returned
}
