% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hamlet}
\alias{hamlet}
\title{Hamlet}
\format{A character string (\code{vector} of length 1).}
\source{
\url{http://www.gutenberg.org/ebooks/1524}
}
\usage{
hamlet
}
\description{
Hamlet by William Shakespeare.
}
\keyword{datasets}
