% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{init_naive_classifer}
\alias{init_naive_classifer}
\alias{init_naive_classifer.factor}
\title{Naive Bayes Classifier}
\usage{
init_naive_classifer(labels)

\method{init_naive_classifer}{factor}(labels)
}
\arguments{
\item{labels}{A vector of labels (factor).}
}
\description{
Naive Bayes classifier for text classification.
}
\examples{
\dontrun{
init_textanalysis()
classes <- factor(c("A", "B"))
model <- init_naive_classifer(classes)
}

}
