% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{inverse_index}
\alias{inverse_index}
\alias{inverse_index.corpus}
\alias{update_inverse_index.corpus}
\title{Inverse Index}
\usage{
inverse_index(corpus)

\method{inverse_index}{corpus}(corpus)

\method{update_inverse_index}{corpus}(corpus)
}
\arguments{
\item{corpus}{A corpus, as returned vy \code{\link{corpus}}.}
}
\description{
If we are interested in a specific term, we often 
want to know which documents in a corpus contain 
that term. The inverse index tells us this and 
therefore provides a simplistic sort of search 
algorithm.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc1 <- string_document("First document.")
doc2 <- string_document("Second document.")

corpus <- corpus(doc1, doc2)

update_inverse_index(corpus)
inverse_index(corpus)
}

}
