% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic.R
\name{lda}
\alias{lda}
\alias{lda.dtm}
\title{Latent Dirichlet Analysis}
\usage{
lda(dtm, topics = 2L, iter = 1000L, alpha = 0.1, beta = 0.1)

\method{lda}{dtm}(dtm, topics = 2L, iter = 1000L, alpha = 0.1,
  beta = 0.1)
}
\arguments{
\item{dtm}{An object of class \code{document_term_matrix} 
as returned by \code{document_term_matrix}.}

\item{topics, iter}{Number of topics and iterations.}

\item{alpha}{Dirichlet dist. hyperparameter for topic distribution per document.
\code{alpha < 1} yields a sparse topic mixture for each document. 
\code{alpha > 1} yields a more uniform topic mixture for each document.}

\item{beta}{Dirichlet dist. hyperparameter for word distribution per topic. 
\code{beta < 1} yields a sparse word mixture for each topic.
\code{beta > 1} yields a more uniform word mixture for each topic.}
}
\value{
A list containing.
\itemize{
 \item{\code{ntopics_nwords} \code{ntopics * nwords} Sparse matrix of probabilities s.t. \eqn{sum(ntopics_nwords, 1) == 1}.}
 \item{\code{ntopics_ndocs} \code{ntopics * ndocs} Dense matrix of probabilities s.t. \eqn{sum(theta, 1) == 1}.}
}
}
\description{
Perform Latent Dirichlet Analysis or lda on a term-document matrix.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc1 <- string_document("First document. Another sentence")
doc2 <- string_document("Some example written here.")
doc3 <- string_document("This is a string document")
doc4 <- string_document("Yet another document for the corpus.")

crps <- corpus(doc1, doc2, doc3, doc4)

update_lexicon(crps)

m <- document_term_matrix(crps)
lda <- lda(m, 2L, 1000L, .1, .1)
}

}
