% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{lexical_diversity}
\alias{lexical_diversity}
\alias{lexical_diversity.corpus}
\title{Lexical Frequency}
\usage{
lexical_diversity(corpus, word)

\method{lexical_diversity}{corpus}(corpus, word)
}
\arguments{
\item{corpus}{A corpus, as returned vy \code{\link{corpus}}.}

\item{word}{Word to check frequency.}
}
\description{
How often a term occurs across all of the documents.
}
