% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{lexical_frequency}
\alias{lexical_frequency}
\alias{lexical_frequency.corpus}
\title{Lexical Frequency}
\usage{
lexical_frequency(corpus, word)

\method{lexical_frequency}{corpus}(corpus, word)
}
\arguments{
\item{corpus}{A corpus, as returned vy \code{\link{corpus}}.}

\item{word}{Word to check frequency.}
}
\description{
How often a term occurs across all of the documents.
}
\examples{
\dontrun{
init_textanalysis()

# build documents
doc1 <- string_document("First document.")
doc2 <- string_document("Second document.")

# build corpus
corpus <- corpus(doc1, doc2)

lexical_frequency(corpus, "document")
}

}
