% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{lexicon}
\alias{lexicon}
\alias{update_lexicon}
\alias{lexicon.corpus}
\alias{update_lexicon.corpus}
\alias{update_inverse_index}
\title{Lexicon}
\usage{
lexicon(corpus)

update_lexicon(corpus)

\method{lexicon}{corpus}(corpus)

\method{update_lexicon}{corpus}(corpus)

update_inverse_index(corpus)
}
\arguments{
\item{corpus}{A corpus, as returned vy \code{\link{corpus}}.}
}
\description{
The lexicon of a corpus consists of all the terms that 
occur in any document in the corpus. The lexical frequency 
of a term tells us how often a term occurs across all of the 
documents. Often the most interesting words in a document 
are those words whose frequency within a document is higher 
than their frequency in the corpus as a whole.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc1 <- string_document("First document.")
doc2 <- string_document("Second document.")

# do not automatically update
corpus <- corpus(doc1, doc2, update_lexicon = FALSE)

update_lexicon(corpus)
lexicon(corpus)
}

}
