% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{lexicon_size}
\alias{lexicon_size}
\alias{lexicon_size.corpus}
\title{Lexicon Size}
\usage{
lexicon_size(corpus)

\method{lexicon_size}{corpus}(corpus)
}
\arguments{
\item{corpus}{A corpus, as returned vy \code{\link{corpus}}.}
}
\description{
Tells the total number of terms in a lexicon.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc1 <- string_document("First document.")
doc2 <- string_document("Second document.")

corpus <- corpus(doc1, doc2)

update_lexicon(corpus)
lexicon_size(corpus)
}

}
