% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic.R
\name{lsa}
\alias{lsa}
\alias{lsa.corpus}
\alias{lsa.dtm}
\title{Latent Semantic Analysis}
\usage{
lsa(corpus)

\method{lsa}{corpus}(corpus)

\method{lsa}{dtm}(corpus)
}
\arguments{
\item{corpus}{A corpus, as returned vy \code{\link{corpus}}.}
}
\description{
Perform Latent Semantic Analysis or lda on a corpus.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc1 <- string_document("First document. Another sentence")
doc2 <- string_document("Some example written here.")
doc3 <- string_document("This is a string document")
doc4 <- string_document("Yet another document for the corpus.")

crps <- corpus(doc1, doc2, doc3, doc4)

lsa(crps)
}

}
