% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Merge the document features, sentiment and topic output, and meta data}
\usage{
merge_data(main_df, list_df, ID = "ID")
}
\arguments{
\item{main_df}{The main data frame (ideally contains all key values)}

\item{list_df}{A list of data frames}

\item{ID}{A character, the key to be merged}
}
\value{
An integrated data frame
}
\description{
Merge many features of documents, to output an integrated data_frame
}
\examples{
merged <- merge_data(main_df = document \%>\% select(ID, LENGTH, DOCUMENT_DATE),
                     list_df = list(stm, meta$city, meta$state),
                     ID = "ID")
}
\author{
Jiacheng He
}
