% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalysis-package.r
\docType{data}
\name{negations}
\alias{negations}
\title{Table of negations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27 rows and 1 columns.}
\usage{
negations
}
\description{
Table of negations
}
\keyword{datasets}
