% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{ngram_complexity}
\alias{ngram_complexity}
\alias{ngram_complexity.ngram_document}
\alias{ngram_complexity.ngram_documents}
\alias{ngram_complexity.JuliaObject}
\title{Determine NGram Complexity}
\usage{
ngram_complexity(document)

\method{ngram_complexity}{ngram_document}(document)

\method{ngram_complexity}{ngram_documents}(document)

\method{ngram_complexity}{JuliaObject}(document)
}
\arguments{
\item{document}{A document as returned by the \code{*_document} 
family of functions, i.e.: \code{\link{string_document}}.}
}
\description{
Determine whether an \code{ngram_document} (output of \code{\link{ngram_document}}) 
contains unigrams, bigrams or a higher-order representation.
}
\examples{
\dontrun{
init_textanalysis()

# build document
unigram <- ngram_document("This is a document.", 1L)
bigram <- ngram_document("This is a document.", 2L)

# test complexity
identical(ngram_complexity(unigram), 1L)
identical(ngram_complexity(bigram), 2L)
}

}
