% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_word_freq.R
\name{plot_word_freq}
\alias{plot_word_freq}
\title{Plot the frequency of the most frequent words}
\usage{
plot_word_freq(df, freq_n = 30, q = NULL, display = FALSE)
}
\arguments{
\item{df}{The document-term data_frame output from create_DTM}

\item{freq_n}{A numeric specifying how many most frequent words we want to display. Defalut to 30.}

\item{q}{A numeric specifying the quantile of tf-idf to remove words. If default to NULL, then don't remove}
}
\value{
A ggplot histogram
}
\description{
Input the document-term data_frame, then plot the the frequency of the most frequent words with histogram
}
\examples{
plot_word_freq(text, freq_n = 30, q = 0.2)

}
\author{
Jiacheng He
}
