% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{predict_class}
\alias{predict_class}
\alias{predict_class.naive_bayes_model}
\title{Predict Class}
\usage{
predict_class(model, data, text)

\method{predict_class}{naive_bayes_model}(model, data, text)
}
\arguments{
\item{model}{A naive Bayes classifier as returned by 
\code{\link{init_naive_classifer}}.}

\item{data}{A data.frame containing \code{text} and \code{labels}.}

\item{text}{Text to train model on \code{labels}.}
}
\description{
Predict class using naive Bayes classifier.
}
\examples{
\dontrun{
init_textanalysis()
classes <- factor(c("financial", "legal"))
model <- init_naive_classifer(classes)

train <- tibble::tibble(
  text = c("this is financial doc", "this is legal doc"),
  labels = factor("financial", "legal")
)

train_naive_classifier(model, train, text, labels)

test <- tibble::tibble(
  text = "this should be predicted as a legal document"
)
predict_class(model, test, text)
}

}
