% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{prepare}
\alias{prepare}
\alias{prepare.document}
\alias{prepare.corpus}
\title{Preprocess Document}
\usage{
prepare(text, ...)

\method{prepare}{document}(text, remove_corrupt_utf8 = TRUE,
  remove_case = TRUE, strip_stopwords = TRUE, strip_numbers = TRUE,
  strip_html_tags = TRUE, strip_punctuation = TRUE,
  remove_words = NULL, strip_non_letters = FALSE,
  strip_spares_terms = FALSE, strip_frequent_terms = FALSE,
  strip_articles = FALSE, strip_indefinite_articles = FALSE,
  strip_definite_articles = FALSE, strip_preposition = FALSE,
  strip_pronouns = FALSE, ...)

\method{prepare}{corpus}(text, remove_corrupt_utf8 = TRUE,
  remove_case = TRUE, strip_stopwords = TRUE, strip_numbers = TRUE,
  strip_html_tags = TRUE, strip_punctuation = TRUE,
  remove_words = NULL, strip_non_letters = FALSE,
  strip_spares_terms = FALSE, strip_frequent_terms = FALSE,
  strip_articles = FALSE, strip_indefinite_articles = FALSE,
  strip_definite_articles = FALSE, strip_preposition = FALSE,
  strip_pronouns = FALSE, ..., update_lexicon = TRUE,
  update_inverse_index = TRUE)
}
\arguments{
\item{text}{An object inheriting of class \code{document} or \code{corpus}.}

\item{...}{Other special classes}

\item{remove_corrupt_utf8}{Remove corrupt UTF8 characters.}

\item{remove_case}{Convert to lowercase.}

\item{strip_stopwords}{Remove stopwords, i.e.: "all", "almost", "alone".}

\item{strip_numbers}{Remove numbers.}

\item{strip_html_tags}{Remove html tags, including the style and script tags.}

\item{strip_punctuation}{Remove punctuation.}

\item{remove_words}{Remove the occurences of words from `doc`.}

\item{strip_non_letters}{Remove anything non-numeric.}

\item{strip_spares_terms}{Remove sparse terms.}

\item{strip_frequent_terms}{Remove frequent terms.}

\item{strip_articles}{Remove articles: "a", "an", "the".}

\item{strip_indefinite_articles}{Removes indefinite articles:  "a", "an".}

\item{strip_definite_articles}{Remove "the".}

\item{strip_preposition}{Remove preprositions, i.e.: "across", "around", "before".}

\item{strip_pronouns}{Remove pronounces, i.e.: "I", "you", "he", "she".}

\item{update_lexicon}{Whether to update the lexicon of the corpus, 
see \code{\link{update_lexicon}}.}

\item{update_inverse_index}{Whether to update the inverse index of the corpus, 
see \code{\link{update_inverse_index}}.}
}
\description{
Preprocess the document, note that this replaces the object in place.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc <- string_document("This <span>is</span> a very short document!.!")

# replaces in place!
prepare(doc)
get_text(doc)
}

}
\seealso{
\code{\link{stem_words}} to stem your document.
}
