% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{prepare_document}
\alias{prepare_document}
\alias{prepare_document.document}
\title{Preprocess Document}
\usage{
prepare_document(document, ...)

\method{prepare_document}{document}(document, remove_corrupt_utf8 = TRUE,
  remove_case = TRUE, strip_stopwords = TRUE, strip_numbers = TRUE,
  strip_html_tags = TRUE, strip_punctuation = TRUE,
  remove_words = NULL, strip_non_letters = FALSE,
  strip_spares_terms = FALSE, strip_frequent_terms = FALSE,
  strip_articles = FALSE, strip_indefinite_articles = FALSE,
  strip_definite_articles = FALSE, strip_preposition = FALSE,
  strip_pronouns = FALSE, ...)
}
\arguments{
\item{document}{A document as returned by the \code{*_document} 
family of functions, i.e.: \code{\link{string_document}}.}

\item{...}{Other special classes}

\item{remove_corrupt_utf8}{Remove corrupt UTF8 characters.}

\item{remove_case}{Convert to lowercase.}

\item{strip_stopwords}{Remove stopwords, i.e.: "all", "almost", "alone".}

\item{strip_numbers}{Remove numbers.}

\item{strip_html_tags}{Remove html tags, including the style and script tags.}

\item{strip_punctuation}{Remove punctuation.}

\item{remove_words}{Remove the occurences of words from `doc`.}

\item{strip_non_letters}{Remove anything non-numeric.}

\item{strip_spares_terms}{Remove sparse terms.}

\item{strip_frequent_terms}{Remove frequent terms.}

\item{strip_articles}{Remove articles: "a", "an", "the".}

\item{strip_indefinite_articles}{Removes indefinite articles:  "a", "an".}

\item{strip_definite_articles}{Remove "the".}

\item{strip_preposition}{Remove preprositions, i.e.: "across", "around", "before".}

\item{strip_pronouns}{Remove pronounces, i.e.: "I", "you", "he", "she".}
}
\description{
Preprocess the document, note that this replaces the object in place.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc <- string_document("This <span>is</span> a very short document!!!")

# replaces in place!
prepare_document(doc)
get_text(doc)
}

}
\seealso{
\code{\link{stem_document}} to stem your document.
}
