% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{push_document}
\alias{push_document}
\alias{push_document.corpus}
\title{Add document}
\usage{
push_document(corpus, document)

\method{push_document}{corpus}(corpus, document)
}
\arguments{
\item{corpus}{A corpus, as returned vy \code{\link{corpus}}.}

\item{document}{An object of class \code{document}.}
}
\description{
Add a document to a corpus.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc1 <- string_document("First document.")
doc2 <- string_document("Second document.")

# length 1
(corpus <- corpus(doc1))
push_document(corpus, doc2)

# length 2
corpus
}

}
