% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_LexisNexis.r
\name{read_LexisNexis}
\alias{read_LexisNexis}
\title{Read LexisNexis data from the database}
\usage{
read_LexisNexis(db, metadata = FALSE, format = FALSE)
}
\arguments{
\item{db}{The name of the LexisNexis database}

\item{metadata}{A logical specifying whether to read the metadata. Default to FALSE}

\item{format}{A logical. If TRUE, then organize the metadata (make it nice to merge).}
}
\value{
A document data_frame; or a list of document data_frame and metadata
}
\description{
Throw in the database and return many data frames
}
\examples{
document <- read_LexisNexis("LexisNexis_v1.db", metadata = FALSE)

LN_Orlando <- read_LexisNexis("LexisNexis_v1.db", metadata = TRUE)
document <- LN_Orlando$document
state <- LN_Orlando$metadata$state
}
\author{
Jiacheng He
}
