% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{remove_corrupt_utf8}
\alias{remove_corrupt_utf8}
\alias{remove_corrupt_utf8.corpus}
\alias{remove_corrupt_utf8.documents}
\alias{remove_corrupt_utf8.document}
\title{Remove Corrupt UTF8}
\usage{
remove_corrupt_utf8(text)

\method{remove_corrupt_utf8}{corpus}(text)

\method{remove_corrupt_utf8}{documents}(text)

\method{remove_corrupt_utf8}{document}(text)
}
\arguments{
\item{text}{A \code{document}, a \code{corpus}, or \code{documents}.}
}
\description{
Remove corrupt UTF8 characters that might cause issues, recommended.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc <- string_document("this document is clean")

# replaces in place!
remove_corrupt_utf8(doc)
}

}
