% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reuters}
\alias{reuters}
\title{Reuters}
\format{A data.frame of 1045 rows and 2 columns
\describe{
  \item{\code{text}}{ text of article}
  \item{\code{category}}{ category, one of 10}
}}
\source{
\url{https://github.com/news-r/nltk4r}
}
\usage{
reuters
}
\description{
Reuters articles on ten different commodities.
}
\keyword{datasets}
