% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_analysis.R
\name{rp_model}
\alias{rp_model}
\alias{rp_model.dtm}
\alias{rp_model.corpus}
\title{Random Projection Model}
\usage{
rp_model(text, ...)

\method{rp_model}{dtm}(text, ...)

\method{rp_model}{corpus}(text, ...)
}
\arguments{
\item{text}{An object inheriting of class \code{document} or \code{corpus}.}

\item{...}{Any other options to pass to the model 
\url{https://zgornel.github.io/StringAnalysis.jl/dev/examples/#Dimensionality-reduction-1}.}
}
\description{
In mathematics and statistics, random projection 
is a technique used to reduce the dimensionality 
of a set of points which lie in Euclidean space. 
Random projection methods are powerful methods 
known for their simplicity and less erroneous 
output compared with other methods. According to 
experimental results, random projection preserve 
distances well, but empirical results are sparse. 
They have been applied to many natural language 
tasks under the name of random indexing. The core 
idea behind random projection is given in the 
Johnson-Lindenstrauss lemma which states that if 
points in a vector space are of sufficiently high 
dimension, then they may be projected into a suitable 
lower-dimensional space in a way which approximately 
preserves the distances between the points.
}
\examples{
\dontrun{
# Use stringanalysis backend!
init_stringanalysis()

# build document
doc1 <- string_document("First document.")
doc2 <- string_document("Second document.")

crps <- corpus(doc1, doc2)
dtm <- document_term_matrix(crps)
model <- rp_model(dtm)
}

}
