% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_LDA.R
\name{run_LDA}
\alias{run_LDA}
\title{Run LDA}
\usage{
run_LDA(df, topic_num, seed_num = 731, topic_term_n = NULL, q = NULL)
}
\arguments{
\item{df}{The document-term data_frame output from create_DTM}

\item{topic_num}{An integer specifying the number of topics}

\item{seed_num}{An integer specifying the random seed in the algorithm}

\item{topic_term_n}{An integer sepcifying how many most frequent words within each topic to display. If default to NULL, then don't plot}

\item{q}{A numeric specifying the quantile of tf-idf to remove words. If default to NULL, then don't remove}
}
\value{
A list of the beat matrix, the gamma matrix, and the topic-term histogram
}
\description{
Run LDA
}
\examples{
run_LDA(text, 6)

lda_output <- run_LDA(text, 6, seed_num = 731, topic_term_n = 20)
print(lda_output$topic_term_plot)
}
\author{
Jiacheng He
}
