% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalysis-package.r
\docType{data}
\name{scriptAllBigrams}
\alias{scriptAllBigrams}
\title{Table of bi-grams from Monty Python script speech lines}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24364 rows and 12 columns.}
\usage{
scriptAllBigrams
}
\description{
Table of bi-grams from Monty Python script speech lines
}
\keyword{datasets}
