% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalysis-package.r
\docType{data}
\name{scriptAllWords}
\alias{scriptAllWords}
\title{Table of words from Monty Python script speech lines}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28349 rows and 12 columns.}
\usage{
scriptAllWords
}
\description{
Table of words from Monty Python script speech lines
}
\keyword{datasets}
