% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalysis-package.r
\docType{data}
\name{scriptBigrams}
\alias{scriptBigrams}
\title{Table of bi-grams from Monty Python script speech lines without stop words}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2523 rows and 12 columns.}
\usage{
scriptBigrams
}
\description{
Table of bi-grams from Monty Python script speech lines without stop words
}
\keyword{datasets}
