% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalysis-package.r
\docType{data}
\name{scriptData}
\alias{scriptData}
\title{Table of Monty Python scripts}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 6 columns.}
\usage{
scriptData
}
\description{
Table of Monty Python scripts
}
\keyword{datasets}
