% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalysis-package.r
\docType{data}
\name{scriptLines}
\alias{scriptLines}
\title{Table of Monty Python script lines}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6292 rows and 10 columns.}
\usage{
scriptLines
}
\description{
Table of Monty Python script lines
}
\keyword{datasets}
