% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalysis-package.r
\docType{data}
\name{scriptSpeech}
\alias{scriptSpeech}
\title{Table of Monty Python script speech lines}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3993 rows and 11 columns.}
\usage{
scriptSpeech
}
\description{
Table of Monty Python script speech lines
}
\keyword{datasets}
