% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalysis-package.r
\docType{data}
\name{scriptURLs}
\alias{scriptURLs}
\title{URLs table for various Monty Python scripts}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 80 rows and 2 columns.}
\usage{
scriptURLs
}
\description{
URLs table for various Monty Python scripts
}
\keyword{datasets}
