% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalysis-package.r
\docType{data}
\name{scriptWords}
\alias{scriptWords}
\title{Table of words from Monty Python script speech lines without stop words}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9262 rows and 12 columns.}
\usage{
scriptWords
}
\description{
Table of words from Monty Python script speech lines without stop words
}
\keyword{datasets}
