% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_key_words.R
\name{search_key_words}
\alias{search_key_words}
\title{Select documents containing a certain set of key words}
\usage{
search_key_words(df, key_words)
}
\arguments{
\item{df}{The document-term data_frame output from create_DTM}

\item{key_words}{A list of key words}
}
\value{
A subset of the input document-term data frame with only the documents containing the key words
}
\description{
Input the document-term data_frame and specify the key words
}
\examples{
search_key_words(text, list("Trump"))
search_key_words(text, list("lgbt", "gunman", "killer"))
}
\author{
Jiacheng He
}
