% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sentiment}
\alias{sentiment}
\alias{sentiment.document}
\alias{sentiment.corpus}
\title{Sentiment Analyzer}
\usage{
sentiment(document)

\method{sentiment}{document}(document)

\method{sentiment}{corpus}(corpus)
}
\description{
Find the sentiment score (between 0 and 1) of a 
word, sentence or a Document.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc <- string_document("An awesome, great, simply brillaint, function!")

sentiment(pos_doc)
}

}
