% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_post_negation.R
\name{sentiment_post_negation}
\alias{sentiment_post_negation}
\title{Extract the sentiment from documents after considering a set of negation words}
\usage{
sentiment_post_negation(df, ID, text, lexicon, negation_words)
}
\arguments{
\item{df}{The original data frame (LexisNexis, Twitter, etc)}

\item{ID}{The name of the variable identifying each document in the original data frame}

\item{text}{The name of the variable containing the texts we want to analyze in the original data frame}

\item{lexicon}{The sentiment database. lexicon = c("affin", "bing", "nrc", "loughran")}

\item{negation_words}{A list of character}
}
\value{
A data_frame where each row represents a document and each column represents a sentiment.
If we use "affin" then it returns the sentiment score
}
\description{
To address the ""NOT HAPPY" problem
}
\examples{
sentiment_post_negation(document, ID, FULL_TEXT, "nrc", list("no", "not"))

}
\author{
Jiacheng He
}
