% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_seed}
\alias{set_seed}
\title{Set Seed in Julia}
\usage{
set_seed(x)
}
\arguments{
\item{x}{Seed, an integer.}
}
\description{
Set a seed in Julia, similar to \code{set.seed} in R.
}
\examples{
\dontrun{set_seed(42)}

}
