% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{stem_words}
\alias{stem_words}
\alias{stem_words.document}
\title{Stem}
\usage{
stem_words(document)

\method{stem_words}{document}(document)
}
\arguments{
\item{document}{A document as returned by the \code{*_document} 
family of functions, i.e.: \code{\link{string_document}}.}
}
\description{
Stem document
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc <- string_document("They write, it writes")

# replaces in place!
stem_words(doc)
get_text(doc)
}

}
