% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{strip_articles}
\alias{strip_articles}
\alias{strip_articles.corpus}
\alias{strip_articles.documents}
\alias{strip_articles.document}
\title{Strip Articles}
\usage{
strip_articles(text)

\method{strip_articles}{corpus}(text)

\method{strip_articles}{documents}(text)

\method{strip_articles}{document}(text)
}
\arguments{
\item{text}{An object inheriting of class \code{document} or \code{corpus}.}
}
\description{
Remove articles, or the words "a", "an", "the".
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc <- string_document("This document has an article.")

# replaces in place!
strip_articles(doc)
get_text(doc)
}

}
