% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{strip_definite_articles}
\alias{strip_definite_articles}
\alias{strip_definite_articles.corpus}
\alias{strip_definite_articles.documents}
\alias{strip_definite_articles.document}
\title{Strip Definite Articles}
\usage{
strip_definite_articles(text)

\method{strip_definite_articles}{corpus}(text)

\method{strip_definite_articles}{documents}(text)

\method{strip_definite_articles}{document}(text)
}
\arguments{
\item{text}{An object inheriting of class \code{document} or \code{corpus}.}
}
\description{
Remove articles, or the word "the".
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc <- string_document("This document has the article.")

# replaces in place!
strip_definite_articles(doc)
get_text(doc)
}

}
