% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{strip_frequent_terms}
\alias{strip_frequent_terms}
\alias{strip_frequent_terms.corpus}
\alias{strip_frequent_terms.documents}
\alias{strip_frequent_terms.document}
\title{Strip Sparse Terms}
\usage{
strip_frequent_terms(text)

\method{strip_frequent_terms}{corpus}(text)

\method{strip_frequent_terms}{documents}(text)

\method{strip_frequent_terms}{document}(text)
}
\arguments{
\item{text}{An object inheriting of class \code{document} or \code{corpus}.}
}
\description{
Remove sparse words.
}
\examples{
\dontrun{
init_textanalysis()

# build document
# must be lowercase
doc <- string_document("This is 1 document.")

# replaces in place!
strip_frequent_terms(doc)
get_text(doc)
}

}
