% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{strip_html_tags}
\alias{strip_html_tags}
\alias{strip_html_tags.corpus}
\alias{strip_html_tags.documents}
\alias{strip_html_tags.document}
\title{Strip HTML Tags}
\usage{
strip_html_tags(text)

\method{strip_html_tags}{corpus}(text)

\method{strip_html_tags}{documents}(text)

\method{strip_html_tags}{document}(text)
}
\arguments{
\item{text}{An object inheriting of class \code{document} or \code{corpus}.}
}
\description{
Remove html tags.
}
\examples{
\dontrun{
init_textanalysis()

# build document
# must be lowercase
doc <- string_document("This is a <span>document</span>.")

# replaces in place!
strip_html_tags(doc)
get_text(doc)
}

}
