% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{strip_indefinite_articles}
\alias{strip_indefinite_articles}
\alias{strip_indefinite_articles.corpus}
\alias{strip_indefinite_articles.documents}
\alias{strip_indefinite_articles.document}
\title{Strip Indefinite Articles}
\usage{
strip_indefinite_articles(text)

\method{strip_indefinite_articles}{corpus}(text)

\method{strip_indefinite_articles}{documents}(text)

\method{strip_indefinite_articles}{document}(text)
}
\arguments{
\item{text}{An object inheriting of class \code{document} or \code{corpus}.}
}
\description{
Remove articles, or the words "a", "an".
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc <- string_document("This is an indifite article.")

# replaces in place!
strip_indefinite_articles(doc)
get_text(doc)
}

}
