% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{strip_non_letters}
\alias{strip_non_letters}
\alias{strip_non_letters.corpus}
\alias{strip_non_letters.documents}
\alias{strip_non_letters.document}
\title{Strip Stopwords}
\usage{
strip_non_letters(text)

\method{strip_non_letters}{corpus}(text)

\method{strip_non_letters}{documents}(text)

\method{strip_non_letters}{document}(text)
}
\arguments{
\item{text}{An object inheriting of class \code{document} or \code{corpus}.}
}
\description{
Remove stopwords, or the words such as "all", "almost", "alone", (and many more).
}
\examples{
\dontrun{
init_textanalysis()

# build document
# must be lowercase
doc <- string_document("This is 1 document.")

# replaces in place!
strip_non_letters(doc)
get_text(doc)
}

}
