% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{strip_numbers}
\alias{strip_numbers}
\alias{strip_numbers.corpus}
\alias{strip_numbers.documents}
\alias{strip_numbers.document}
\title{Strip Numbers}
\usage{
strip_numbers(text)

\method{strip_numbers}{corpus}(text)

\method{strip_numbers}{documents}(text)

\method{strip_numbers}{document}(text)
}
\arguments{
\item{text}{An object inheriting of class \code{document} or \code{corpus}.}
}
\description{
Remove numbers.
}
\examples{
\dontrun{
init_textanalysis()

# build document
# must be lowercase
doc <- string_document("This is 1 document.")

# replaces in place!
strip_numbers(doc)
get_text(doc)
}

}
