% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{strip_preprositions}
\alias{strip_preprositions}
\alias{strip_preprositions.corpus}
\alias{strip_preprositions.documents}
\alias{strip_preprositions.document}
\title{Strip Preprositions}
\usage{
strip_preprositions(text)

\method{strip_preprositions}{corpus}(text)

\method{strip_preprositions}{documents}(text)

\method{strip_preprositions}{document}(text)
}
\arguments{
\item{text}{An object inheriting of class \code{document} or \code{corpus}.}
}
\description{
Remove preprositions, or the words such as "across", "around", "before" (and many more).
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc <- string_document("A preprosition is in the document.")

# replaces in place!
strip_preprositions(doc)
get_text(doc)
}

}
