% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{strip_pronouns}
\alias{strip_pronouns}
\alias{strip_pronouns.corpus}
\alias{strip_pronouns.documents}
\alias{strip_pronouns.document}
\title{Strip Pronouns}
\usage{
strip_pronouns(text)

\method{strip_pronouns}{corpus}(text)

\method{strip_pronouns}{documents}(text)

\method{strip_pronouns}{document}(text)
}
\arguments{
\item{text}{An object inheriting of class \code{document} or \code{corpus}.}
}
\description{
Remove pronouns, or the words such as  "I", "you", "he", "she", (and many more).
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc <- string_document("She is the pronoun.")

# replaces in place!
strip_pronouns(doc)
get_text(doc)
}

}
