% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{strip_punctuation}
\alias{strip_punctuation}
\alias{strip_punctuation.corpus}
\alias{strip_punctuation.documents}
\alias{strip_punctuation.document}
\title{Strip Punctuation}
\usage{
strip_punctuation(text)

\method{strip_punctuation}{corpus}(text)

\method{strip_punctuation}{documents}(text)

\method{strip_punctuation}{document}(text)
}
\arguments{
\item{text}{An object inheriting of class \code{document} or \code{corpus}.}
}
\description{
Remove punctuation marks.
}
\examples{
\dontrun{
init_textanalysis()

# build document
# must be lowercase
doc <- string_document(".This is a document!!")

# replaces in place!
strip_punctuation(doc)
get_text(doc)
}

}
