% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{strip_stopwords}
\alias{strip_stopwords}
\alias{strip_stopwords.corpus}
\alias{strip_stopwords.documents}
\alias{strip_stopwords.document}
\title{Strip Stopwords}
\usage{
strip_stopwords(text)

\method{strip_stopwords}{corpus}(text)

\method{strip_stopwords}{documents}(text)

\method{strip_stopwords}{document}(text)
}
\arguments{
\item{text}{An object inheriting of class \code{document} or \code{corpus}.}
}
\description{
Remove stopwords, or the words such as "all", "almost", "alone", (and many more).
}
\examples{
\dontrun{
init_textanalysis()

# build document
# must be lowercase
doc <- string_document("The document includes stop words.")

# replaces in place!
strip_stopwords(doc)
get_text(doc)
}

}
