% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{summarize}
\alias{summarize}
\alias{summarize.document}
\alias{summarize.corpus}
\title{Summarize}
\usage{
summarize(text, ns = 2L)

\method{summarize}{document}(text, ns = 2L)

\method{summarize}{corpus}(text, ns = 2L)
}
\arguments{
\item{text}{An object of class \code{document} or \code{corpus}.}

\item{ns}{Number of sentences.}
}
\description{
Simple text-rank based summarizer.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc <- string_document(
  paste("Assume this Short Document as an example.",
    "Assume this as an example summarizer.", 
    "This has too few sentences."
  )
)

summarize(doc, ns = 2L)
}

}
