% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextAnalysis-package.r
\docType{package}
\name{TextAnalysis}
\alias{TextAnalysis}
\alias{TextAnalysis-package}
\title{A helper package for Lock Data Text Analysis
training.}
\description{
A helper package for Lock Data Text Analysis
training.
}
