% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{tf}
\alias{tf}
\alias{tf.dtm}
\alias{tf.JuliaObject}
\title{Term Frequency}
\usage{
tf(dtm)

\method{tf}{dtm}(dtm)

\method{tf}{JuliaObject}(dtm)
}
\arguments{
\item{dtm}{An object of class \code{document_term_matrix} 
as returned by \code{document_term_matrix}.}
}
\description{
Compute term-frequency.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc1 <- string_document("First document.")
doc2 <- string_document("Second document.")

corpus <- corpus(doc1, doc2)

update_lexicon(corpus)
m <- document_term_matrix(corpus)
tf(m)
}

}
