% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{tf_idf}
\alias{tf_idf}
\alias{tf_idf.dtm}
\alias{tf_idf.JuliaObject}
\title{Term Frequency Inverse Document Freqency}
\usage{
tf_idf(dtm)

\method{tf_idf}{dtm}(dtm)

\method{tf_idf}{JuliaObject}(dtm)
}
\arguments{
\item{dtm}{An object of class \code{document_term_matrix} 
as returned by \code{document_term_matrix}.}
}
\description{
Compute term-frequency inverse-document-frequency.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc1 <- string_document("First document.")
doc2 <- string_document("Second document.")

corpus <- corpus(doc1, doc2)

update_lexicon(corpus)
m <- document_term_matrix(corpus)
tf_idf(m)
}

}
