% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{to_documents}
\alias{to_documents}
\alias{to_documents.character}
\alias{to_documents.data.frame}
\title{Create Multiple Documents}
\usage{
to_documents(documents, ...)

\method{to_documents}{character}(documents, ..., type = c("string",
  "token", "ngram"))

\method{to_documents}{data.frame}(documents, ..., text, title = NULL,
  language = NULL, author = NULL, timestamp = NULL,
  type = c("string", "token", "ngram"))
}
\arguments{
\item{documents}{A \code{data.frame}, or \code{vector} 
containing documents.}

\item{...}{Not actually being used.}

\item{type}{Type of \code{*_document} function to use, see \code{\link{documents}}.}

\item{text}{Bare column name containing data.}

\item{title, language, author, timestamp}{Bare column name 
to metadata to add to documents, see \code{\link{document_metadata}} .}
}
\value{
An object of class \code{documents}.
}
\description{
Create multiple documents from a \code{data.frame}, 
or \code{vector} of characters.
}
\examples{
\dontrun{
init_textanalysis()

docs <- c(
  "This is a document.",
  "This is another document."
)
to_documents(docs)

docs_df <- tibble::tibble(
  txt = docs,
  title = c("A", "B") 
)
to_documents(docs_df)
title_(docs_df)
}

}
\seealso{
\code{\link{documents}} to read a single document.
}
