% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{train_naive_classifier}
\alias{train_naive_classifier}
\alias{train_naive_classifier.naive_bayes_model}
\title{Train Naive Bayes Classifier}
\usage{
train_naive_classifier(model, data, text, labels)

\method{train_naive_classifier}{naive_bayes_model}(model, data, text,
  labels)
}
\arguments{
\item{model}{A naive Bayes classifier as returned by 
\code{\link{init_naive_classifer}}.}

\item{data}{A data.frame containing \code{text} and \code{labels}.}

\item{text}{Text to train model on \code{labels}.}

\item{labels}{Bare column name containing labels.}
}
\description{
Train Naive Bayes classifier (\code{\link{init_naive_classifer}}).
}
\examples{
\dontrun{
init_textanalysis()
classes <- factor(c("financial", "legal"))
model <- init_naive_classifer(classes)

train <- tibble::tibble(
  text = c("this is financial doc", "this is legal doc"),
  labels = factor("financial", "legal")
)

train_naive_classifier(model, train, text, labels)
}

}
