% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_LDA.R
\name{tune_LDA}
\alias{tune_LDA}
\title{Find the optimal number of topics}
\usage{
tune_LDA(df, k_range, seed_num = 731, mc.cores = 4L, verbose = TRUE,
  q = NULL)
}
\arguments{
\item{df}{The document-term data_frame output from create_DTM}

\item{k_range}{A vector of integers specifying the range of number of topics}

\item{seed_num}{An integer specifying the random seed in the algorithm}

\item{mc.cores}{An integer specifying how many cores to use for computing. Default to 4}

\item{verbose}{A logical specifying whether to display the progress when the algorith is running. Default to TRUE}

\item{q}{A numeric specifying the quantile of tf-idf to remove words. If default to NULL, then don't remove}
}
\value{
A list of matrix reporting the criterion measure for each number,
the optimal number based on Griffiths2004,
the optimal number based on CaoJuan2009,
the optimal number based on Arun2010,
the optimal number based on Deveaud2014,
and the plot of the measure matrix
}
\description{
Tune a specified range of numbers, find the optimal number of topics
based on some criterion measures ("Griffiths2004", "CaoJuan2009", "Arun2010", "Deveaud2014")
}
\examples{
tune_result <- tune_LDA(text, k_range = 2:15, q = 0.1)
print(tune_result$k_plot)
}
\author{
Jiacheng He
}
