% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_LDA_eyeball.R
\name{tune_LDA_eyeball}
\alias{tune_LDA_eyeball}
\title{Use your eyeball to detect the optimal number of topics which deliever meaningful stuff}
\usage{
tune_LDA_eyeball(df, k_range, seed_num = 731, topic_term_n, q = NULL)
}
\arguments{
\item{df}{The document-term data_frame output from create_DTM}

\item{k_range}{A vector of integers specifying the range of number of topics}

\item{seed_num}{An integer specifying the random seed in the algorithm}

\item{topic_term_n}{An integer sepcifying how many most frequent words within each topic to display}

\item{q}{A numeric specifying the quantile of tf-idf to remove words. If default to NULL, then don't remove}
}
\value{
A list of Topic-term histograms
}
\description{
Input the range of number of topics then it output the topic-term histogram
from the training result of each number
}
\examples{

lda_output <- run_LDA(text, 6, seed_num = 731, topic_term_n = 20)
print(lda_output[[1]])
print(lda_output[[2]])
}
\author{
Jiacheng He
}
