% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{corpus}
\alias{corpus}
\alias{corpus.document}
\alias{corpus.list}
\alias{directory_corpus}
\title{Corpus}
\usage{
corpus(doc, ...)

\method{corpus}{document}(doc, ...)

\method{corpus}{list}(doc, ...)

directory_corpus(directory)
}
\arguments{
\item{doc}{First \code{document}, a \code{list}, or a \code{vector} of documents.}

\item{...}{Objects inheriting of class \code{document} to build a corpus.}

\item{directory}{Path to a directory of text files.}
}
\description{
Build a corpus from documents or a directory of text files.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc1 <- string_document("First document.")
doc2 <- string_document("Second document.")

corpus <- corpus(doc1, doc2)
}

}
