% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{corpus}
\alias{corpus}
\alias{corpus.document}
\alias{corpus.documents}
\alias{directory_corpus}
\title{Corpus}
\usage{
corpus(document, ..., update_lexicon = TRUE,
  update_inverse_index = TRUE)

\method{corpus}{document}(document, ..., update_lexicon = TRUE,
  update_inverse_index = TRUE)

\method{corpus}{documents}(document, ..., update_lexicon = TRUE,
  update_inverse_index = TRUE)

directory_corpus(directory, update_lexicon = TRUE,
  update_inverse_index = TRUE)
}
\arguments{
\item{document}{First \code{document}, a \code{list}, or a \code{vector} of documents.}

\item{...}{Objects inheriting of class \code{document} to build a corpus.}

\item{update_lexicon}{Whether to update the lexicon, see \code{\link{update_lexicon}}.}

\item{update_inverse_index}{Whether to update the inverse index, see \code{\link{update_inverse_index}}.}

\item{directory}{Path to a directory of text files.}
}
\description{
Build a corpus from documents or a directory of text files.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc1 <- string_document("First document.")
doc2 <- string_document("Second document.")

corpus <- corpus(doc1, doc2)
}

}
