% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{create_hash_function}
\alias{create_hash_function}
\title{Hash Trick}
\usage{
create_hash_function(cardinality)
}
\arguments{
\item{cardinality}{Max index used for hashing (default 100).}
}
\description{
Hash Trick
}
\examples{
\dontrun{
init_textanalysis()

hash_func <- create_hash_function(doc, ns = 2L)
}

}
