% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{documents}
\alias{documents}
\alias{file_document}
\alias{string_document}
\alias{token_document}
\alias{ngram_document}
\title{Document}
\usage{
file_document(path)

string_document(text)

token_document(text)

ngram_document(text, ...)
}
\arguments{
\item{path}{The path to the file.}

\item{text}{The text as a character string, or tokens, or ngrams as a list.}

\item{...}{Other positonal arguments.}
}
\value{
An object of class \code{document}.
}
\description{
The basic unit of text analysis is a document. 
The textanalysis package allows one to work with documents 
stored in a variety of formats.
}
\examples{
\dontrun{
init_textanalysis()

doc <- "This is a document."
string_document(doc)
ngram_document(doc, 2L)
}

}
\seealso{
\code{\link{directory_corpus}} to read a dicrectory of files as corpus, 
and \code{\link{to_documents}} to parse a vector or data.frame to documents.
}
