% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScriptLines.R
\name{getScriptLines}
\alias{getScriptLines}
\title{Get Monty Python script speech lines}
\usage{
getScriptLines()
}
\value{
data.frame of speech lines from Monty Python scripts
}
\description{
Get Monty Python script speech lines
}
\examples{
head(getScriptLines())
}
