% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{init}
\alias{init}
\alias{init_textanalysis}
\alias{init_stats}
\alias{setup_textanalysis}
\title{Initialise Session}
\usage{
init_textanalysis(...)

init_stats()

setup_textanalysis(...)
}
\arguments{
\item{...}{Arguments passed to \link[JuliaCall]{julia_setup}.}
}
\description{
Initialise a session, installs TextAnalysis Juia dependency if needed.
}
\section{Packages}{

Packages installed by \code{init_textanalysis} are:
\itemize{
  \item{\code{TextAnalysis}}
  \item{\code{Languages}}
}
Packages installed by \code{init_stats} are:
\itemize{
  \item{\code{MultivariateStats}}
  \item{\code{Clustering}}
}
}

