% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{init}
\alias{init}
\alias{init_textanalysis}
\alias{setup_textanalysis}
\title{Initialise Session}
\usage{
init_textanalysis(...)

setup_textanalysis(...)
}
\arguments{
\item{...}{Arguments passed to \link[JuliaCall]{julia_setup}.}
}
\description{
Initialise a session, installs TextAnalysis Juia dependency if needed.
}
