% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{init}
\alias{init}
\alias{init_textanalysis}
\alias{setup_textanalysis}
\title{Initialise Session}
\usage{
init_textanalysis(..., type = c("stable", "latest"))

setup_textanalysis(..., type = c("stable", "latest"))
}
\arguments{
\item{...}{Arguments passed to \link[JuliaCall]{julia_setup}.}

\item{type}{Whether to install the stable version from the 
registry (recommended) or the latest version from github.}
}
\description{
Initialise a session, installs TextAnalysis Juia dependency if needed.
}
\section{Packages}{

Packages installed by \code{init_textanalysis} are:
\itemize{
  \item{\code{TextAnalysis}}
  \item{\code{Languages}}
}
}

