% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{remove_case}
\alias{remove_case}
\alias{remove_case.corpus}
\alias{remove_case.documents}
\alias{remove_case.document}
\title{Remove Upper case}
\usage{
remove_case(text)

\method{remove_case}{corpus}(text)

\method{remove_case}{documents}(text)

\method{remove_case}{document}(text)
}
\arguments{
\item{text}{An object inheriting of class \code{document} or \code{corpus}.}
}
\description{
Turn everything to lowercase.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc <- string_document("ThIs DoCuMENT HAs UppERCase")

# replaces in place!
remove_case(doc)
get_text(doc)
}

}
