% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sentiment}
\alias{sentiment}
\alias{sentiment.document}
\alias{sentiment.corpus}
\title{Sentiment Analyzer}
\usage{
sentiment(text)

\method{sentiment}{document}(text)

\method{sentiment}{corpus}(text)
}
\arguments{
\item{text}{An object of class \code{document} or \code{corpus}.}
}
\description{
Find the sentiment score (between 0 and 1) of a 
word, sentence or a Document based on a trained 
model (using Flux) on IMDB word corpus with weights 
saved are used to calculate the sentiments.
}
\examples{
\dontrun{
init_textanalysis()

# build document
doc <- string_document(
  "An awesome, great, simply brillaint, function!"
)

sentiment(doc)
}

}
